# specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Semiont API",
    "version": "0.1.0",
    "description": "Semantic Knowledge Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API Server"
    }
  ],
  "components": {
    "schemas": {
...
      "Motivation": {
        "type": "string",
        "enum": [
          "assessing",
          "bookmarking",
          "classifying",
          "commenting",
          "describing",
          "editing",
          "highlighting",
          "identifying",
          "linking",
          "moderating",
          "questioning",
          "replying",
          "tagging"
        ],
        "description": "W3C Web Annotation motivation vocabulary - https://www.w3.org/TR/annotation-vocab/#motivation"
      },
...
      "GetResourceResponse": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceDescriptor"
          },
          "annotations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            },
            "description": "All annotations for the resource (highlights, references, assessments, etc.)"
          },
          "entityReferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Annotation"
            },
            "description": "Annotations that reference this resource from other resources"
          }
        },
        "required": [
          "resource",
          "annotations",
          "entityReferences"
        ]
      },
...
      "UpdateResourceRequest": {
        "type": "object",
        "properties": {
          "entityTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "ReferenceLLMContextResponse": {
        "type": "object",
        "properties": {
          "reference": {
            "$ref": "#/components/schemas/Annotation"
          },
          "sourceResource": {
            "$ref": "#/components/schemas/ResourceDescriptor"
          },
          "targetResource": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceDescriptor"
              },
              {
                "type": "null"
              }
            ]
          },
          "sourceContext": {
            "type": "object",
            "properties": {
              "before": {
                "type": "string"
              },
              "selected": {
                "type": "string"
              },
              "after": {
                "type": "string"
              }
            },
            "required": [
              "before",
              "selected",
              "after"
            ]
          },
          "targetContext": {
            "type": "object",
            "properties": {
              "content": {
                "type": "string"
              },
              "summary": {
                "type": "string"
              }
            },
            "required": [
              "content"
            ]
          },
          "suggestedResolution": {
            "type": "object",
            "properties": {
              "resourceId": {
                "type": "string"
              },
              "resourceName": {
                "type": "string"
              },
              "confidence": {
                "type": "number"
              },
              "reasoning": {
                "type": "string"
              }
            },
            "required": [
              "resourceId",
              "resourceName",
              "confidence",
              "reasoning"
            ]
          }
        },
        "required": [
          "reference",
          "sourceResource"
        ]
      },
...
      "Agent": {
        "type": "object",
        "description": "Minimal Person/Organization for attribution (W3C PROV compatible)",
        "additionalProperties": true,
        "properties": {
          "@id": {
            "type": "string",
            "format": "uri",
            "description": "Unique identifier for the agent"
          },
          "@type": {
            "description": "Type(s) of agent (Person, Organization, Software)",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "minItems": 1
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_sha1": {
            "type": "string"
          },
          "homepage": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "ResourceDescriptor": {
        "type": "object",
        "description": "Metadata about a resource (1:1 with its URI). JSON-LD subject is @id. Link to concrete bytes via representations.",
        "required": [
          "@context",
          "@id",
          "name",
          "representations"
        ],
        "additionalProperties": true,
        "properties": {
          "@context": {
            "description": "JSON-LD context; URI, object, or array of these.",
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "type": "string",
                      "format": "uri"
                    },
                    {
                      "type": "object",
                      "additionalProperties": true
                    }
                  ]
                }
              }
            ]
          },
          "@id": {
            "type": "string",
            "format": "uri",
            "description": "Canonical URI/URN of the resource being described."
          },
          "@type": {
            "description": "Type(s) of the resource (IRIs/CURIEs via @context).",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "minItems": 1
              }
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "identifier": {
            "description": "Persistent identifiers (e.g., DOI, URN).",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "minItems": 1
              },
              {
                "type": "object",
                "properties": {
                  "@id": {
                    "type": "string",
                    "format": "uri"
                  },
                  "value": {
                    "type": "string"
                  },
                  "scheme": {
                    "type": "string"
                  }
                },
                "additionalProperties": true
              }
            ]
          },
          "about": {
            "description": "Topics (IRIs or strings).",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "minItems": 1
              }
            ]
          },
          "sameAs": {
            "description": "Equivalent/authoritative references.",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "isPartOf": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "hasPart": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "license": {
            "type": "string",
            "format": "uri"
          },
          "version": {
            "type": "string"
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "dateModified": {
            "type": "string",
            "format": "date-time"
          },
          "wasDerivedFrom": {
            "description": "W3C PROV - source resources this was derived from",
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            ]
          },
          "wasAttributedTo": {
            "description": "W3C PROV - agents responsible for this resource",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Agent"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            ]
          },
          "conformsTo": {
            "description": "Profile/shape URI this resource description conforms to.",
            "oneOf": [
              {
                "type": "string",
                "format": "uri"
              },
              {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uri"
                }
              }
            ]
          },
          "availableFormats": {
            "description": "Convenience set summarizing media types across representations.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "representations": {
            "description": "Managed or referenced byte-level renditions of this resource.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Representation"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Representation"
                }
              }
            ]
          },
          "archived": {
            "type": "boolean",
            "description": "Application-specific: Whether this resource is archived"
          },
          "entityTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Application-specific: Entity types for this resource"
          },
          "isDraft": {
            "type": "boolean",
            "description": "Application-specific: Whether this resource is a draft"
          },
          "creationMethod": {
            "type": "string",
            "enum": [
              "api",
              "upload",
              "ui",
              "reference",
              "clone",
              "generated"
            ],
            "description": "Application-specific: How this resource was created"
          },
          "sourceAnnotationId": {
            "type": "string",
            "description": "Application-specific: ID of annotation that triggered generation"
          },
          "sourceResourceId": {
            "type": "string",
            "description": "Application-specific: ID of source resource for clones/derivatives"
          }
        }
      },
...
      "Annotation": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "string",
            "const": "http://www.w3.org/ns/anno.jsonld",
            "default": "http://www.w3.org/ns/anno.jsonld",
            "description": "W3C Web Annotation JSON-LD context"
          },
          "type": {
            "type": "string",
            "const": "Annotation",
            "default": "Annotation",
            "description": "W3C Annotation type"
          },
          "id": {
            "type": "string"
          },
          "motivation": {
            "$ref": "#/components/schemas/Motivation"
          },
          "target": {
            "oneOf": [
              {
                "type": "string",
                "description": "IRI of the entire resource being annotated"
              },
              {
                "$ref": "#/components/schemas/AnnotationTarget"
              }
            ],
            "description": "W3C Web Annotation target - can be a simple IRI string (entire resource) or an object with source and optional selector (fragment)"
          },
          "body": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationBody"
                },
                "maxItems": 0,
                "description": "Empty array for stub references (Phase 1 pattern)"
              },
              {
                "$ref": "#/components/schemas/AnnotationBody",
                "description": "Single body (TextualBody or SpecificResource)"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AnnotationBody"
                },
                "description": "Array of mixed TextualBody (tagging) and SpecificResource (linking) bodies"
              }
            ],
            "description": "W3C Web Annotation body - can be empty array, single body, or array of bodies"
          },
          "creator": {
            "$ref": "#/components/schemas/Agent"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          },
          "generator": {
            "$ref": "#/components/schemas/Agent"
          }
        },
        "required": [
          "@context",
          "type",
          "id",
          "motivation",
          "target",
          "body"
        ]
      },
...
      "AnnotationTarget": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "description": "IRI of the resource being annotated"
          },
          "selector": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TextPositionSelector"
              },
              {
                "$ref": "#/components/schemas/TextQuoteSelector"
              },
              {
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TextPositionSelector"
                    },
                    {
                      "$ref": "#/components/schemas/TextQuoteSelector"
                    }
                  ]
                }
              }
            ],
            "description": "Optional selector to identify a specific segment of the source resource"
          }
        },
        "required": [
          "source"
        ],
        "description": "W3C Web Annotation target object - source is required, selector is optional"
      },
...
    },
...
  },
  "paths": {
    "/resources/{id}": {
      "get": {
        "summary": "Get Resource by URI (W3C Content Negotiation)",
        "description": "W3C-compliant globally resolvable resource URI with content negotiation. Returns JSON-LD metadata (Accept: application/ld+json), raw representation (Accept: text/plain, text/markdown, etc.), or redirects to frontend (?view=semiont). Requires authentication.",
        "tags": [
          "W3C URIs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource ID"
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["semiont"]
            },
            "description": "Optional view parameter. If set to 'semiont', redirects to frontend viewer."
          }
        ],
        "responses": {
          "200": {
            "description": "Resource metadata (JSON-LD) or raw representation",
            "content": {
              "application/ld+json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResourceResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "302": {
            "description": "Redirect to frontend viewer (when ?view=semiont)"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "patch": {
        "summary": "Update Resource",
        "description": "Update resource metadata (append-only operations - name and content are immutable)",
        "tags": [
          "W3C URIs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resource updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetResourceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
...
    "/api/resources/{resourceId}/references/{referenceId}/llm-context": {
      "get": {
        "summary": "Get Reference LLM Context",
        "description": "Get reference with full context for LLM processing",
        "tags": [
          "Resources",
          "AI"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "resourceId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "referenceId",
            "in": "path"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "true"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "false"
                  ]
                },
                {
                  "type": "boolean"
                }
              ],
              "default": true
            },
            "required": false,
            "name": "includeSourceContext",
            "in": "query"
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "true"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "false"
                  ]
                },
                {
                  "type": "boolean"
                }
              ],
              "default": true
            },
            "required": false,
            "name": "includeTargetContext",
            "in": "query"
          },
          {
            "schema": {
              "type": "number",
              "minimum": 100,
              "maximum": 5000,
              "default": 1000
            },
            "required": false,
            "name": "contextWindow",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Reference LLM context",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceLLMContextResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}